agregar validaciones

+++ backend/routers/poligonos.js
@@ -97,28 +97,32 @@ routerPoligonos.get('/:id_poligono', validaIdPoligono, async (req, res) => {
   }
 });
 
-//delete a poligono
+//delete a poligono, 
 
 routerPoligonos.delete('/:id_poligono', validaIdPoligono, async (req, res) => {
   try {
-    const { id_poligono } = req.params;
-    
-    const  operacion  = req.method;
-    const  id_usuarioAuditoria =req.headers['id_usuario'];
+      const { id_poligono } = req.params;
+      
+      //if the polygon id route parameter is empty "" throw error message
+      if (!id_poligono) {
+          return res.status(400).json({ error: 'El campo id_poligono es requerido' });
+      }
+      
+      const operacion = req.method;
+      const id_usuarioAuditoria = req.headers['id_usuario'];
+      
+      const eliminarPoligono = await pool.query('DELETE FROM poligonos WHERE id_poligono = $1', [id_poligono]);
 
-    const eliminarPoligono = await pool.query('DELETE FROM poligonos WHERE id_poligono = $1', [id_poligono]);
+      if (eliminarPoligono.rowCount === 0) {
+          return res.status(404).json({ error: 'Polígono no encontrado' });
+      }
 
-    if (eliminarPoligono.rowCount === 0) {
-      // No se eliminó ningún polígono, ya que no se encontró en la base de datos
-      return res.status(404).json({ error: 'Polígono no encontrado' });
-    }
+      auditar(operacion, id_usuarioAuditoria);
 
-    auditar(operacion,id_usuarioAuditoria);
-
-    res.json('Polígono eliminado');
+      res.json('Polígono eliminado');
   } catch (error) {
-    console.log(error);
-    res.status(500).json({ error: 'Error al eliminar el polígono' });
+      console.log(error);
+      res.status(500).json({ error: 'Error al eliminar el polígono' });
   }
 });
